takes care of consumptions

+++ src/main/java/parc/model/concrete/Consumption.java
@@ -48,7 +48,7 @@ public class Consumption {
         return distance;
     }
 
-    public void setDistance(int distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
